#280 Delete FC_REFLECT_TEMPLATE from votes

See merge request PBSA/peerplays!76

+++ tests/cli/son.cpp
@@ -263,8 +263,8 @@ BOOST_AUTO_TEST_CASE( son_voting )
       auto nathan_votes = con.wallet_api_ptr->get_votes("nathan");
       BOOST_REQUIRE(nathan_votes.votes_for_sons);
       BOOST_CHECK_EQUAL(nathan_votes.votes_for_sons->size(), 2);
-      BOOST_CHECK_EQUAL((uint32_t)nathan_votes.votes_for_sons->at(0).id.instance, son1_obj.id.instance());
-      BOOST_CHECK_EQUAL((uint32_t)nathan_votes.votes_for_sons->at(1).id.instance, son2_obj.id.instance());
+      BOOST_CHECK_EQUAL(nathan_votes.votes_for_sons->at(0).id.instance(), son1_obj.id.instance());
+      BOOST_CHECK_EQUAL(nathan_votes.votes_for_sons->at(1).id.instance(), son2_obj.id.instance());
 
       // Withdraw vote for a son1account
       BOOST_TEST_MESSAGE("Withdraw vote for a son1account");
@@ -285,7 +285,7 @@ BOOST_AUTO_TEST_CASE( son_voting )
       nathan_votes = con.wallet_api_ptr->get_votes("nathan");
       BOOST_REQUIRE(nathan_votes.votes_for_sons);
       BOOST_CHECK_EQUAL(nathan_votes.votes_for_sons->size(), 1);
-      BOOST_CHECK_EQUAL((uint32_t)nathan_votes.votes_for_sons->at(0).id.instance, son2_obj.id.instance());
+      BOOST_CHECK_EQUAL(nathan_votes.votes_for_sons->at(0).id.instance(), son2_obj.id.instance());
 
       // Withdraw vote for a son2account
       BOOST_TEST_MESSAGE("Withdraw vote for a son2account");
